Rename parser to nodeparser to avoid std module name conflict
